scripts/update-gh-pages: manage a Helm repo
Make the update-gh-pages.sh script to maintain a Helm charts repository
under charts/ subdirectory in gh-pages. The script now (always) scans
throught all release assets and injects any found Helm chart archives
into the Helm repo. In practice, new assets in all Github releases are
scanned and the Helm repo is updated on any update of the master or
release branches or on any new tags. Asset ids are tracked/cached in
order to avoid unnecessary downloads, but also, to capture any changes
in assets that were already merged in the repo index.

After this a user is able to do something like

$ helm repo add nfd http://kubernetes-sigs.github.io/node-feature-discovery/charts
  ...
$ helm repo update
  ...
$ helm install nfd/node-feature-discovery --namespace nfd --create-namespace --generate-name
  ...