source: make sources register themselves
Implement new registration infrastructure under source/. This change
makes label sources and the nfd-worker core more loosely coupled, making
it easier to refactor and move the code around.

Also, create a separate interface for sources that are configurable in
order to eliminate boilerplate code.

Add safety checks to the sources that they actually implement the
interfaces they should.

Change all methods of the sources to use pointer receivers, in the sake
of consistency and predictability (of behavior).