source/kernel: implement FeatureSource
Separate feature discovery and creation of feature labels in the kernel
source.

Move kernelutils from source/internal back to the source/kernel package.

Move the implementation of the Kconfig custom rule to the source/kernel
package as well. This simplifies the code as the rule matcher is able to
directly use data discovered by the kernel source.

Also, introduces a new BoolFeatureValue type to indicate whether the
value of a bool is actually valid. This is needed to handle detection
failures of bool features.