source/kernel: implement FeatureSource
Separate feature discovery and creation of feature labels in the kernel
source.

Move kernelutils from source/internal back to the source/kernel package.
Change the cpuid kconfig rule to utilize GetFeatures of the kernel
source.

Also, introduces a new BoolFeatureValue type to indicate whether the
value of a bool is actually valid. This is needed to handle detection
failures of bool features.