source/custom: add generic per-domain rules
Implement new generic rule processing that covers all feature
sources (that implement the FeatureSource interface). This relies on the
unified data model provided by the FeatureSource interface as well as
the generic rule processing that was added to the source package.

With this patch any new features added will be automatically available
for custom rules, without any additional work. Rule hierarchy follows
the source/feature hierarchy by design.

The new rules are available under two new fields in the custom rule
specification, namely 'matchAny' and 'matchAll'. MatchAny
effectively implements a logical OR between the rulesets (similar to
matchOn), i.e. a feature is found if any of the rulesets match.
Correspondingly, matchAll implements a logical AND, i.e. a match for all
of the rulesets is required.

The old rules (cpuID, kConfig, loadedKMod, nodename, pciID, usbID) are
only available under the 'matchOn' field whereas the new rules are only
available under the 'matchAny' and 'matchAll' fields.

On the top level, a logical AND is applied over matchOn, matchAny and
matchAll. That is, all of these fields (if specified) must match.

The configuration format for the new rules is

  {matchAny | matchAll}:
    <domain>.<feature>:
      <attribute>:
        op: <operation>
        value:
        - <list-of-values>
        - ...

The configuration also supports some "shortforms" for convenience:

   <domain-a>.<feature-a>: [<attr-1>, <attr-2>=<val-2>]
   <domain-b>.<feature-b>:
     <attr-3>:
     <attr-4>: <val-4>

is equal to

   <domain-a>.<feature-a>:
     <attr-1>: {op: Exists}
     <attr-2>: {op: In, value: [<val-2>]}
   <domain-b>.<feature-b>:
     <attr-3>: {op: Exists}
     <attr-4>: {op: In, value: [<val-4>]}

That is:

  kernel.config: ["X86", "INIT_ENV_ARG_LIMIT=32"]
  pci.device:
    vendor: "8086"

is the same as

  kernel.config:
    X86: {op: Exists}
    INIT_ENV_ARG_LIMIT: {op: In, values: ["32"]}
  pci.device:
    vendor: {op: In, value: ["8086"]

Some configuration examples below. In order to match a CPUID feature the
following snippet can be used:

  - name: cpu-test-1
    matchAny:
    - cpu.cpuid:
        AESNI: {op: Exists
        AVX: {op: Exists}

In order to match against a loaded kernel module and OS version:

  - name: kernel-test-1
    matchAny:
    - kernel,loadedmodule:
        e1000: {op: Exists}
      system.osrelease:
        NAME: {op: InRegexp, values: ["^openSUSE"]}
        VERSION_ID.major: {op: Gt, values: ["14"]}

In order to require a cpu feature, kernel module and one of two specific
PCI devices (taking use of the shortform notation):

  - name: multi-device-test
    matchAll:
    - kernel.loadedmodule: [driver-module]
      cpu.cpuid: [AVX512F]
    matchAny:
    - pci.device:
        vendor: "8086"
        device: "1234"
    - pci.device:
        vendor: "8086"
        device: "abcd"

In order to require a kernel module and both of two specific PCI devices:

  - name: multi-device-test
    matchAll:
    - kernel.loadedmodule:
        driver-module: {op: Exists}
    - pci.device:
        vendor: "8086"
        device: "1234"
    - pci.device:
        vendor: "8086"
        device: "abcd"