source/custom: add generic per-domain rules
Implement new generic rule processing that covers all feature
sources (that implement the FeatureSource interface). This relies on the
unified data model provided by the FeatureSource interface as well as
the generic rule processing that was added to the source package.

With this patch any new features added will be automatically available
for custom rules, without any additional work. Rule hierarchy follows
the source/feature hierarchy by design.

The new rules are available under two new fields in the custom rule
specification, namely 'matchAny' and 'matchAll'. MatchAny
effectively implements a logical OR between the rulesets (similar to
matchOn), i.e. a feature is found if any of the rulesets match.
Correspondingly, matchAll implements a logical AND, i.e. a match for all
of the rulesets is required.

The old rules (cpuID, kConfig, loadedKMod, nodename, pciID, usbID) are
only available under the 'matchOn' field whereas the new rules are only
available under the 'matchAny' and 'matchAll' fields.

A logical AND is applied over matchOn, matchAny and matchAll. That is,
all of these fields (which are specified) must match.

For example, in order to match a CPUID feature the following snippet can
be used:

  - name: cpu-test-1
    matchAny:
    - cpu:
        cpuid: {op: Exists, values: ["AVX", "AESNI"]}

In order to match against a loaded kernel module and OS version:

  - name: kernel-test-1
    matchAny:
    - kernel:
        loadedModule : {op: Exists, values ["e1000"]}
      system:
        osRelease:
          NAME: {op: InRegexp, values: ["^openSUSE"]}
          VERSION_ID.major: {op: Gt, values: ["14"]}

In order to require a cpu feature, kernel module and one of two specific
PCI devices:

  - name: multi-device-test
    matchAll:
    - kernel:
        loadedModule : {op: Exists, values [driver-module]}
      cpu:
        cpuid: {op: Exists, values: ["AVX512F"]}
    matchAny:
    - pci:
        device:
          vendor: "8086"
          device: "1234"
    - pci:
        device:
          vendor: "8086"
          device: "abcd"

In order to require a kernel module and both of two specific PCI devices:

  - name: multi-device-test
    matchAll:
    - kernel:
        loadedModule : {op: Exists, values [driver-module]}
    - pci:
        device:
          vendor: "8086"
          device: "1234"
    - pci:
        device:
          vendor: "8086"
          device: "abcd"