source/custom: add generic per-domain rules
Implement new generic rule processing that covers all feature
sources (that implement the FeatureSource interface). This relies on the
unified data model provided by the FeatureSource interface as well as
the generic rule processing that was added to the source package.

With this patch any new features added will be automatically available
for custom rules, without any additional work. Rule hierarchy follows
the source/feature hierarchy by design.

For example, in order to match a CPUID feature the following snippet can
be used:

  - name: cpu-test-1
    matchOn:
    - cpu:
        cpuid: {op: Exists, values: ["AVX", "AESNI"]}

In order to match against a loaded kernel module and OS version:

  - name: kernel-test-1
    matchOn:
    - kernel:
         loadedModule : {op: Exists, values ["e1000"]}
    - system:
        osRelease:
          NAME: {op: InRegexp, values: ["^openSUSE"]}
          VERSION_ID.major: {op: Gt, values: ["14"]}