source/custom: support label templating
Support templating in custom label names.

First, this patch adds the possibility to specify multiple labels per
rule. They just need to be separated by newlines. An example:

  - name: |
      label-1
      label-2

Second, this patch adds the possibility to specify the label value in
the name field as well. If specified, this will override anything set in
the 'value' field. Examples:

  - name: |
      label-1
      label-2=override-value
    value: default-value

  - name: label-3=custom-value

Third, this patch adds support for string templating with the golang
"text/template" package.  With the new functionality stated above
(multiple labels per name and value overrides with '=') it is possible
to generate multiple dynamic labels per rule. All the matched features
of 'matchAny' and 'matchAll' directives are available for templating in
a nested domain.feature map structure. Matches from legacy rules under
the 'matchOn' directive are not available for templating.

NOTE: The template is executed against each individual rule under every
matchAny and matchAll directive separately and the eventual set of
feature labels is a combination of all these expansions. Consider the
following:

  - name: <template>
    matchOn:
    - <rule#0>
    matchAny:
    - <rule#1>
    - <rule#2>
    matchAll:
    - <rule#3>
    - <rule#4>

In the example above (assuming the overall result is match) the template
would be executed on rule#1 and/or rule#2 (depending on whether both or
only one of them match), then on rule#3 and finally on rule#4 and the
end result would be a union of all these separate expansions.

Some example configurations:

  - name: "{{range .pci.device}}intel-{{ .class}}-{{ .device }}=present\n{{end}}"
    matchAny:
    - pci.device:
        class: {op: InRegexp, value: ["^06"]}
        vendor: ["8086"]

  - name: "{{range .system.osrelease}}system-{{.Name}}={{.Value}}\n{{end}}"
    matchAny:
    - system.osRelease:
        ID: {op: Exists}
        VERSION_ID.major: {op: Exists}

Imaginative template pipelines are possible, of course, but care must be
taken in order to produce understandable and maintainable rule sets.