source/custom: support label templating
Support templating in custom label names.

First, this patch adds the possibility to specify multiple labels per
rule. They just need to be separated by newlines. An example:

  - name: |
      label-1
      label-2

Second, this patch adds the possibility to specify the label value in
the name field as well. If specified, this will override anything set in
the 'value' field. Examples:

  - name: |
      label-1
      label-2=override-value
    value: default-value

  - name: label-3=custom-value

Third, this patch adds support for string templating with the golang
"text/template" package.  With the new functionality stated above
(multiple labels per name and value overrides with '=') it is possible
to generate multiple dynamic labels per rule.  All the matched features
are available for templating in a nested domain.feature map structure.
Some examples:

  - name: "{{range .pci.device}}intel-{{ .class}}-{{ .device }}=present\n{{end}}"
    matchOn:
    - pci:
        device:
          class: {op: InRegexp, value: ["^06"]}
          vendor: ["8086"]

  - name: "{{range .system.osrelease}}system-{{.Name}}={{.Value}}\n{{end}}"
    matchOn:
      - system:
          osRelease:
            ID: {op: Exists}
            VERSION_ID.major: {op: Exists}

Imaginative template pipelines are possible, of course, but care must be
taken in order to produce