pkg/apis/nfd: specify CRD for custom rules
Utilize controller-gen (from sigs.k8s.io/controller-tools) for
generating the CRD and deepcopy methods. Code can be (re-)generated with
"make apigen". Install controller-gen with:

  go install sigs.k8s.io/controller-tools/cmd/controller-gen@v0.6.0

This patch virtually replicates the source.custom.DomainsRule in the CRD
api. LegacyRule is not specified in the CRD api in order to keep it
simple and clean.