Add noLabel field to rule spec and support backreferences
Support backreferencing of output values from previous rules. Enables
complex setups where custom features are further combined together to
form even more sophisticated higher level labels. The output of
preceding rules are available as a special 'rule.matched' feature in
matchAny and matchAll directives. If referencing rules accross multiple
configs/CRDs care must be taken with the ordering.

Processing order of rules in nfd-worker:
1. Static rules
2. Files from /etc/kubernetes/node-feature-discovery/custom.d/
   in alphabetical order. Subdirectories are processed by reading their
   files in alphabetical order.
3. Custom rules from main nfd-worker.conf

In nfd-master the CRs are processed in alphabetical order (based on
their metadata.name).

Also, add 'noLabel' boolean field to the rulespec. Enabling this causes
the rule to not generate any node labels. This may be desired in schemes
where the output of certain rules is only used as an intermediate
variable for other rules.

An example setup:

  - name: kernel-feature
    matchAny:
    - kernel.version:
        major: {op: Gt, value: ["4"]}

  - name: nolabel-feature
    noLabel: true
    matchAll:
    - cpu.cpuid:
        AVX512F: {op: Exists}
      pci.device:
        vendor: {op: In, value: ["8086"]}
        device: {op: In, value: ["1234", "1235"]}

  - name: top-level-feature
    matchAll:
    - rule.matched:
        kernel-feature: "true"
        nolabel-feature: "true"