Stop implicitly adding default prefix to names
This change stops NFD from automatically adding the
"feature.node.kubernetes.io/" prefix to labels, annotations and extended
resources. Taints did not have this property. The user-visible part of
the change is that NodeFeatureRules, local feature files, hooks and
configuration of the "custom" may need to be altered.

The goal of this change is to simplify the operation, making it more
straightforward and easier to understand and debug (kind of WYSIWYG),
eliminating peculiar corner cases. For example:

1. Makes validation simpler and unambiguous
2. Removes the mapping of two values to the same actual node label. E.g.
   previously something like

      labels:
        feature.node.kubernetes.io/foo: bar
        foo: baz

   Could actually result in node label:

     feature.node.kubernetes.io/foo: baz

3. Makes the processing/usagee of the "rule.matched" and "local.labels"
   feature in NodeFeatureRules unambiguous and more understadable. E.g.
   previously you could have node label
   "feature.node.kubernetes.io/local-foo: bar" but in the NodeFeatureRule
   you'd need to use the unprefixed name "local-foo" or the fully
   prefixed name, depending on what was specified in the feature file (or
   hook) on the node(s).

NOTE: This is a breaking change for users who rely on automatic
prefixing with the default feature.node.kubernetes.io/ namespace.
NodeFeatureRules, feature files, hooks and custom rules (configuration
of the "custom" source of nfd-worker) will need to be altered.
Unprefixed labels, annoations and extended resources will be denied by
nfd-master.